pyproject.toml
[project]
name = "da2"
version = "0.1.0"
description = "Depth Anything v2 Small + ONNX Runtime (CUDA) toy project"
requires-python = ">=3.10,<3.13"
dependencies = [
    "onnxruntime-gpu>=1.18.0",
    "huggingface_hub>=0.24.0",
    "numpy>=1.24",
    "pillow>=10.0",
    "opencv-python>=4.8",
    "tqdm>=4.66",
    "click>=8.1",
    "requests>=2.31",
]

[project.scripts]
da2 = "da2.cli:app"

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  { include = "da2", from = "src" }
]

[tool.hatch.build.targets.sdist]
include = [
  "src",
  "README.md",
]

[tool.uv]
index-url = "https://pypi.org/simple"
